% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.springer.R
\name{print.springer}
\alias{print.springer}
\title{print a springer result}
\usage{
\method{print}{springer}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{springer result}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a springer result
}
\seealso{
\code{\link{springer}}
}
