% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{This function changes the format of the longitudinal data from wide format to long format}
\usage{
reformat(k, y, x)
}
\arguments{
\item{k}{the number of repeated measurements/time points.}

\item{y}{the longitudinal response.}

\item{x}{a matrix of predictors, consisting of clinical covariates, genetic and environment factors, as well as gene-environment interactions.}
}
\description{
This function changes the format of the longitudinal data from wide format to long format
}
